hydrate projects and homeFireworks
