+++ src/clientApp.js
@@ -3,6 +3,8 @@ import { hydrateCopyright, hydrateGoToTopButton } from "./clientSections/footer.
 import { hydrateExperienceTime } from "./clientSections/experience.js";
 import { go } from "./util/navUtil.js";
 import { fetchSvgs, hydrateSvgs } from "./clientSections/hydrateSvgs.js";
+import { hydrateProjects } from "./clientSections/projects.js";
+import { hydrateHomeFireworks } from "./clientSections/fireworks-gravity.js";
 
 // Access build-time parameters passed from esbuild define
 const buildConfig = {
@@ -28,6 +30,8 @@ export const initApp = () => {
   hydrateGoToTopButton();
   hydrateExperienceTime();
   hydrateCopyright();
+  hydrateProjects();
+  hydrateHomeFireworks();
 };
 
 initApp();
